refactor(locks): move the postMessage bridge to useLocksAuthFlow.utils, dissolving libs/locks (#2026)
